rename doc.go to gen.go for better name convention
