+++ utils/idempotent/gen.go
@@ -17,6 +17,4 @@
 // shutdown) and defines the core lifecycle SPI. The core lifecycle SPI
 // must be implemented by each plugin.
 
-package dataplane
-
-//go:generate bash -c "protoc -I . dataplane.proto --go_out=plugins=grpc:. --proto_path=$GOPATH/src/ --proto_path=$GOPATH/pkg/mod/  --proto_path=$( go list -f '{{ .Dir }}' -m github.com/golang/protobuf )"
+package idempotent
